refactor(netlink): Generalize socket for multi-protocol support
The existing `netlink.rs` was specific to `NETLINK_ROUTE`. To prepare for adding `NETLINK_NETFILTER` support for conntrack, this commit refactors the netlink socket implementation to be generic.

- A new generic `Socket<P>` is introduced in `src/network/netlink.rs` to handle common send/receive logic.
- All routing-specific code, types, and functions are moved to a new `src/network/netlink_route.rs`, which now uses `Socket<NetlinkRoute>`.
- All imports and type signatures across the codebase have been updated to use this new structure.

This is a pure refactoring with no functional changes.
