+++ src/network/bridge.rs
@@ -3,6 +3,10 @@ use std::{collections::HashMap, fs, net::IpAddr, os::fd::BorrowedFd};
 use crate::dns::aardvark::SafeString;
 use crate::network::core_utils::get_default_route_interface;
 use crate::network::dhcp::{dhcp_teardown, get_dhcp_lease};
+use crate::network::netlink::Socket;
+use crate::network::netlink_route::{
+    parse_create_link_options, CreateLinkOptions, LinkID, NetlinkRoute,
+};
 use crate::{
     dns::aardvark::AardvarkEntry,
     error::{ErrorWrap, NetavarkError, NetavarkErrorList, NetavarkResult},
@@ -34,7 +38,7 @@ use super::{
         IPAMAddresses, IsolateOption, PortForwardConfig, SetupNetwork, TearDownNetwork,
         TeardownPortForward,
     },
-    netlink, sysctl,
+    sysctl,
     types::StatusBlock,
 };
 
@@ -158,7 +162,7 @@ impl driver::NetworkDriver for Bridge<'_> {
 
     fn setup(
         &self,
-        netlink_sockets: (&mut netlink::Socket, &mut netlink::Socket),
+        netlink_sockets: (&mut Socket<NetlinkRoute>, &mut Socket<NetlinkRoute>),
     ) -> NetavarkResult<(StatusBlock, Option<AardvarkEntry<'_>>)> {
         let data = match &self.data {
             Some(d) => d,
@@ -348,7 +352,7 @@ impl driver::NetworkDriver for Bridge<'_> {
 
     fn teardown(
         &self,
-        netlink_sockets: (&mut netlink::Socket, &mut netlink::Socket),
+        netlink_sockets: (&mut Socket<NetlinkRoute>, &mut Socket<NetlinkRoute>),
     ) -> NetavarkResult<()> {
         let mode: Option<String> = parse_option(&self.info.network.options, OPTION_MODE)?;
         let mode = get_bridge_mode_from_string(mode.as_deref())?;
@@ -592,8 +596,8 @@ const IPV6_FORWARD: &str = "net/ipv6/conf/all/forwarding";
 
 /// returns the container veth mac address
 fn create_interfaces(
-    host: &mut netlink::Socket,
-    netns: &mut netlink::Socket,
+    host: &mut Socket<NetlinkRoute>,
+    netns: &mut Socket<NetlinkRoute>,
     data: &InternalData,
     internal: bool,
     rootless: bool,
@@ -601,9 +605,9 @@ fn create_interfaces(
     netns_fd: BorrowedFd<'_>,
 ) -> NetavarkResult<CreateInterfacesResult> {
     let mut sysctl_writer = None;
-    let (bridge_index, mtu, mac) = match host.get_link(netlink::LinkID::Name(
-        data.bridge_interface_name.to_string(),
-    )) {
+    let (bridge_index, mtu, mac) = match host
+        .get_link(LinkID::Name(data.bridge_interface_name.to_string()))
+    {
         Ok(bridge) => {
             let (bridge_index, mtu) = validate_bridge_link(
                 bridge,
@@ -696,7 +700,7 @@ fn create_interfaces(
                 // writer must be create before the bridge is created
                 let sw = sysctl::SysctlDWriter::new(path, sysctls);
 
-                let mut create_link_opts = netlink::CreateLinkOptions::new(
+                let mut create_link_opts = CreateLinkOptions::new(
                     data.bridge_interface_name.to_string(),
                     InfoKind::Bridge,
                 );
@@ -724,7 +728,7 @@ fn create_interfaces(
                 }
 
                 if let Some(vrf_name) = &data.vrf {
-                    let vrf = match host.get_link(netlink::LinkID::Name(vrf_name.to_string())) {
+                    let vrf = match host.get_link(LinkID::Name(vrf_name.to_string())) {
                         Ok(vrf) => check_link_is_vrf(vrf, vrf_name)?,
                         Err(err) => return Err(err).wrap("get vrf to set up bridge interface"),
                     };
@@ -738,9 +742,7 @@ fn create_interfaces(
                 sysctl_writer = Some(sw);
 
                 let link = host
-                    .get_link(netlink::LinkID::Name(
-                        data.bridge_interface_name.to_string(),
-                    ))
+                    .get_link(LinkID::Name(data.bridge_interface_name.to_string()))
                     .wrap("get bridge interface")?;
 
                 let mut mac = None;
@@ -761,7 +763,7 @@ fn create_interfaces(
                         .wrap("add ip addr to bridge")?;
                 }
 
-                host.set_up(netlink::LinkID::ID(link.header.index))
+                host.set_up(LinkID::ID(link.header.index))
                     .wrap("set bridge up")?;
 
                 (link.header.index, mtu, mac)
@@ -791,8 +793,8 @@ fn create_interfaces(
 /// return the container veth mac address
 #[allow(clippy::too_many_arguments)]
 fn create_veth_pair<'fd>(
-    host: &mut netlink::Socket,
-    netns: &mut netlink::Socket,
+    host: &mut Socket<NetlinkRoute>,
+    netns: &mut Socket<NetlinkRoute>,
     data: &InternalData,
     primary_index: u32,
     bridge_mac: Option<Vec<u8>>,
@@ -802,16 +804,15 @@ fn create_veth_pair<'fd>(
     mtu: u32,
 ) -> NetavarkResult<String> {
     let mut peer_opts =
-        netlink::CreateLinkOptions::new(data.container_interface_name.to_string(), InfoKind::Veth);
+        CreateLinkOptions::new(data.container_interface_name.to_string(), InfoKind::Veth);
     peer_opts.mac = data.mac_address.clone().unwrap_or_default();
     peer_opts.mtu = mtu;
     peer_opts.netns = Some(netns_fd);
 
     let mut peer = LinkMessage::default();
-    netlink::parse_create_link_options(&mut peer, peer_opts);
+    parse_create_link_options(&mut peer, peer_opts);
 
-    let mut host_veth =
-        netlink::CreateLinkOptions::new(data.host_interface_name.clone(), InfoKind::Veth);
+    let mut host_veth = CreateLinkOptions::new(data.host_interface_name.clone(), InfoKind::Veth);
     host_veth.mtu = mtu;
     host_veth.primary_index = primary_index;
     host_veth.info_data = Some(InfoData::Veth(InfoVeth::Peer(peer)));
@@ -835,9 +836,7 @@ fn create_veth_pair<'fd>(
     })?;
 
     let veth = netns
-        .get_link(netlink::LinkID::Name(
-            data.container_interface_name.to_string(),
-        ))
+        .get_link(LinkID::Name(data.container_interface_name.to_string()))
         .wrap("get container veth")?;
 
     let mut mac = String::from("");
@@ -890,7 +889,7 @@ fn create_veth_pair<'fd>(
         })?;
 
         if data.ipam.ipv6_enabled {
-            let host_veth = host.get_link(netlink::LinkID::ID(host_link))?;
+            let host_veth = host.get_link(LinkID::ID(host_link))?;
 
             for nla in host_veth.attributes.into_iter() {
                 if let LinkAttribute::IfName(name) = nla {
@@ -903,7 +902,7 @@ fn create_veth_pair<'fd>(
         }
     }
 
-    host.set_up(netlink::LinkID::ID(host_link))
+    host.set_up(LinkID::ID(host_link))
         .wrap("failed to set host veth up")?;
 
     // Ok this is extremely strange, by default the kernel will always choose the mac address with the
@@ -916,7 +915,7 @@ fn create_veth_pair<'fd>(
     // connected otherwise no connectivity is possible at all and I have no idea why but CNI does it
     // also in the same way.
     if let Some(m) = bridge_mac {
-        host.set_mac_address(netlink::LinkID::ID(primary_index), m)
+        host.set_mac_address(LinkID::ID(primary_index), m)
             .wrap("set static mac on bridge")?;
     }
 
@@ -927,7 +926,7 @@ fn create_veth_pair<'fd>(
     }
 
     netns
-        .set_up(netlink::LinkID::ID(veth.header.index))
+        .set_up(LinkID::ID(veth.header.index))
         .wrap("set container veth up")?;
 
     if !internal && !data.no_default_route {
@@ -948,7 +947,7 @@ fn create_veth_pair<'fd>(
 fn validate_bridge_link(
     msg: LinkMessage,
     vlan: bool,
-    netlink: &mut netlink::Socket,
+    netlink: &mut Socket<NetlinkRoute>,
     br_name: &str,
 ) -> NetavarkResult<(u32, u32)> {
     let mut mtu: u32 = 0;
@@ -1038,20 +1037,20 @@ fn check_link_is_vrf(msg: LinkMessage, vrf_name: &str) -> NetavarkResult<LinkMes
 }
 
 fn remove_link(
-    host: &mut netlink::Socket,
-    netns: &mut netlink::Socket,
+    host: &mut Socket<NetlinkRoute>,
+    netns: &mut Socket<NetlinkRoute>,
     mode: BridgeMode,
     br_name: &str,
     container_veth_name: &str,
 ) -> NetavarkResult<bool> {
     netns
-        .del_link(netlink::LinkID::Name(container_veth_name.to_string()))
+        .del_link(LinkID::Name(container_veth_name.to_string()))
         .wrap(format!(
             "failed to delete container veth {container_veth_name}"
         ))?;
 
     let br = host
-        .get_link(netlink::LinkID::Name(br_name.to_string()))
+        .get_link(LinkID::Name(br_name.to_string()))
         .wrap("failed to get bridge interface")?;
 
     let links = host
@@ -1061,7 +1060,7 @@ fn remove_link(
     if links.is_empty() {
         if let BridgeMode::Managed = mode {
             log::info!("removing bridge {br_name}");
-            host.del_link(netlink::LinkID::ID(br.header.index))
+            host.del_link(LinkID::ID(br.header.index))
                 .wrap(format!("failed to delete bridge {container_veth_name}"))?;
             return Ok(true);
         }
