+++ src/network/core_utils.rs
@@ -18,6 +18,8 @@ use std::path::Path;
 use std::str::FromStr;
 
 use super::netlink;
+use crate::network::netlink::Socket;
+use crate::network::netlink_route::{LinkID, NetlinkRoute, Route};
 
 use netlink_packet_route::link::LinkAttribute;
 
@@ -126,7 +128,7 @@ pub fn get_ipam_addresses<'a>(
                 });
             }
 
-            let routes: Vec<netlink::Route> = match create_route_list(&network.routes) {
+            let routes: Vec<Route> = match create_route_list(&network.routes) {
                 Ok(r) => r,
                 Err(e) => {
                     return Err(e);
@@ -274,7 +276,7 @@ pub struct NamespaceOptions {
     /// Note we have to return the File object since the fd is only valid
     /// as long as the File object is valid
     pub file: File,
-    pub netlink: netlink::Socket,
+    pub netlink: Socket<NetlinkRoute>,
 }
 
 pub fn open_netlink_sockets(
@@ -283,11 +285,11 @@ pub fn open_netlink_sockets(
     let netns = open_netlink_socket(netns_path).wrap("open container netns")?;
     let hostns = open_netlink_socket("/proc/self/ns/net").wrap("open host netns")?;
 
-    let host_socket = netlink::Socket::new().wrap("host netlink socket")?;
+    let host_socket = netlink::Socket::<NetlinkRoute>::new().wrap("host netlink socket")?;
     let netns_sock = exec_netns!(
         hostns.as_fd(),
         netns.as_fd(),
-        netlink::Socket::new().wrap("netns netlink socket")
+        netlink::Socket::<NetlinkRoute>::new().wrap("netns netlink socket")
     )?;
 
     Ok((
@@ -307,7 +309,7 @@ fn open_netlink_socket(netns_path: &str) -> NetavarkResult<File> {
 }
 
 pub fn add_default_routes(
-    sock: &mut netlink::Socket,
+    sock: &mut Socket<NetlinkRoute>,
     gws: &[ipnet::IpNet],
     metric: Option<u32>,
 ) -> NetavarkResult<()> {
@@ -321,7 +323,7 @@ pub fn add_default_routes(
                 }
                 ipv4 = true;
 
-                netlink::Route::Ipv4 {
+                Route::Ipv4 {
                     dest: ipnet::Ipv4Net::new(Ipv4Addr::new(0, 0, 0, 0), 0)?,
                     gw: v4.addr(),
                     metric,
@@ -333,7 +335,7 @@ pub fn add_default_routes(
                 }
                 ipv6 = true;
 
-                netlink::Route::Ipv6 {
+                Route::Ipv6 {
                     dest: ipnet::Ipv6Net::new(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0), 0)?,
                     gw: v6.addr(),
                     metric,
@@ -346,9 +348,7 @@ pub fn add_default_routes(
     Ok(())
 }
 
-pub fn create_route_list(
-    routes: &Option<Vec<types::Route>>,
-) -> NetavarkResult<Vec<netlink::Route>> {
+pub fn create_route_list(routes: &Option<Vec<types::Route>>) -> NetavarkResult<Vec<Route>> {
     match routes {
         Some(rs) => rs
             .iter()
@@ -357,12 +357,12 @@ pub fn create_route_list(
                 let dst = r.destination;
                 let mtr = r.metric;
                 match (gw, dst) {
-                    (IpAddr::V4(gw4), IpNet::V4(dst4)) => Ok(netlink::Route::Ipv4 {
+                    (IpAddr::V4(gw4), IpNet::V4(dst4)) => Ok(Route::Ipv4 {
                         dest: dst4,
                         gw: gw4,
                         metric: mtr,
                     }),
-                    (IpAddr::V6(gw6), IpNet::V6(dst6)) => Ok(netlink::Route::Ipv6 {
+                    (IpAddr::V6(gw6), IpNet::V6(dst6)) => Ok(Route::Ipv6 {
                         dest: dst6,
                         gw: gw6,
                         metric: mtr,
@@ -398,7 +398,7 @@ pub fn is_using_systemd() -> bool {
 }
 
 /// Returns the *first* interface with a default route or an error if no default route interface exists.
-pub fn get_default_route_interface(host: &mut netlink::Socket) -> NetavarkResult<LinkMessage> {
+pub fn get_default_route_interface(host: &mut Socket<NetlinkRoute>) -> NetavarkResult<LinkMessage> {
     let routes = host.dump_routes().wrap("dump routes")?;
 
     for route in routes {
@@ -416,7 +416,7 @@ pub fn get_default_route_interface(host: &mut netlink::Socket) -> NetavarkResult
         // if there is no dest we have a default route
         // return the output interface for this route
         if !dest && out_if > 0 {
-            return host.get_link(netlink::LinkID::ID(out_if));
+            return host.get_link(LinkID::ID(out_if));
         }
     }
     Err(NetavarkError::msg("failed to get default route interface"))
