+++ src/network/dhcp.rs
@@ -7,8 +7,10 @@ use std::str::FromStr;
 use crate::dhcp_proxy::lib::g_rpc::NetworkConfig;
 use crate::dhcp_proxy::proxy_conf::DEFAULT_UDS_PATH;
 
+use super::core_utils;
 use super::driver::DriverInfo;
-use super::{core_utils, netlink};
+use crate::network::netlink::Socket;
+use crate::network::netlink_route::{LinkID, NetlinkRoute};
 
 pub type DhcpLeaseInfo = (Vec<NetAddress>, Option<Vec<IpAddr>>, Option<Vec<String>>);
 
@@ -160,14 +162,14 @@ pub fn release_dhcp_lease(
     Ok(())
 }
 
-pub fn dhcp_teardown(info: &DriverInfo, sock: &mut netlink::Socket) -> NetavarkResult<()> {
+pub fn dhcp_teardown(info: &DriverInfo, sock: &mut Socket<NetlinkRoute>) -> NetavarkResult<()> {
     let ipam = core_utils::get_ipam_addresses(info.per_network_opts, info.network)?;
     let if_name = info.per_network_opts.interface_name.clone();
 
     // If we are using DHCP, we need to at least call to the proxy so that
     // the proxy's cache can get updated and the current lease can be released.
     if ipam.dhcp_enabled {
-        let dev = sock.get_link(netlink::LinkID::Name(if_name)).wrap(format!(
+        let dev = sock.get_link(LinkID::Name(if_name)).wrap(format!(
             "get container interface {}",
             &info.per_network_opts.interface_name
         ))?;
